Proofreading edits
Ch. 25: Adding Latin-language tag to another word
Ch. 25: Capitalizing "Oh no" as a separate sentence after an exclamation
Ch. 25: Formatting "Ruth Elton," contextually explained as a novel, as a book with full typing metadata
Fixing the italicization of the title of the cover image in the Colophon
Fixing a word-separation typo in the Endnotes
Ch. 28: Adding a broken rule in the newspaper headlines section to prevent indention